@@ -1,11 +1,11 @@
 package com.doug.projects.transitdelayservice.controller;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
 
-@Controller
+@RestController
 @CrossOrigin(origins = {"http://localhost:3000", "www.my-precious-time.com"})
 public class HealthController {
     @GetMapping("/health")
