@@ -35,23 +35,28 @@ public class DelayWriterCronService {
 
     @Scheduled(fixedRate = 300000)
     public void getDelayAndWriteToDb() {
-        if (doesCronRun != null && !doesCronRun) return;
-        //get realtime delays from metro
-        RealtimeTransitResponse transitResponse = realtimeMetroService.getCurrentRunData();
-        //remove basic fields for timestamp checking / iteration
-        if (nullTransitFields(transitResponse)) {
-            log.info("transitResponse:{} data invalid, will not write anything", transitResponse);
-            return;
-        }
-        log.info("Completed realtimeRequest successfully, building model...");
-        List<RouteTimestamp> routeTimestamps = adaptor.convertFrom(transitResponse);
-        List<RouteTimestamp> uniqueRouteTimestamps = DelayWriterCronService.removeDuplicates(routeTimestamps);
-        log.info("Built model successfully from {} objects, yielding {} values for database write.",
-                transitResponse.getEntity().size(), uniqueRouteTimestamps.size());
-        if (!routeTimestampRepository.writeRouteTimestamps(uniqueRouteTimestamps)) {
-            log.error("try again next time... failed writing routeTimestamps :/");
+        try {
+            if (doesCronRun != null && !doesCronRun)
+                return;
+            //get realtime delays from metro
+            RealtimeTransitResponse transitResponse = realtimeMetroService.getCurrentRunData();
+            //remove basic fields for timestamp checking / iteration
+            if (nullTransitFields(transitResponse)) {
+                log.info("transitResponse:{} data invalid, will not write anything", transitResponse);
+                return;
+            }
+            log.info("Completed realtimeRequest successfully, building model...");
+            List<RouteTimestamp> routeTimestamps = adaptor.convertFrom(transitResponse);
+            List<RouteTimestamp> uniqueRouteTimestamps = DelayWriterCronService.removeDuplicates(routeTimestamps);
+            log.info("Built model successfully from {} objects, yielding {} values for database write.",
+                    transitResponse.getEntity()
+                    .size(), uniqueRouteTimestamps.size());
+            if (!routeTimestampRepository.writeRouteTimestamps(uniqueRouteTimestamps)) {
+                throw new RuntimeException("Failed to write route timestamps");
+            }
+        } catch (Exception e) {
+            log.error("Failed to write route timestamps", e);
         }
     }
-
 
 }
